+++ src/components/PageTitle.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import type { Snippet } from 'svelte';
+
+	const { children }: { children: Snippet<[]> } = $props();
+</script>
+
+<h1 class="text-2xl font-bold">{@render children()}</h1>
