+++ src/routes/(app)/+layout.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	const { children } = $props();
+</script>
+
+<main class="container mx-auto p-4">
+	{@render children()}
+</main>
