+++ src/routes/(app)/account/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import PageTitle from '@components/PageTitle.svelte';
+</script>
+
+<PageTitle>Accounts</PageTitle>
