+++ src/routes/+page.svelte
@@ -1,2 +1,4 @@
-<h1>Welcome to SvelteKit</h1>
-<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
+<main class="container mx-auto py-8">
+	<h1 class="text-4xl font-bold">Welcome to SV-Money</h1>
+	<p class="text-xl">This is personal money manager.</p>
+</main>
