+++ src/routes/Header.svelte
@@ -1,9 +1,10 @@
 <header class="h-[60px] bg-slate-600 px-8">
-	<div id="left" class="flex h-full w-full items-center gap-4">
+	<div id="left" class="flex h-full w-full items-center gap-6">
 		<div id="logo" class="font-logo text-3xl font-bold text-white">
 			<a href="/">SVMoney</a>
 		</div>
-		<menu class="text-white">
+		<menu class="flex gap-4 text-white">
+			<li><a href="/account">Account</a></li>
 			<li><a href="/showroom">Showroom</a></li>
 		</menu>
 	</div>
